go: Initial commit
Common packages for working on Go projects.

I frequently use this from my editor with:

- https://github.com/flox/flox-direnv

And the following in `.envrc`:

    use flox --remote floxenvs/go

In future the editor dependencies could be broken out to a separate
`go-editor` or `go-lsp` environment.

The `GOENV` hook is copied from our `init --auto-setup` hook with the
exception of:

1. Path is modified from `goenv` to `go` to match the other example
   environments in this repo
2. `go get .` is omitted because it's surprising to fetch dependencies
   when activating an environment and we have plans to remove it from
   the default manifest.

The example module which includes a third-party module to prove that
dependency fetching still works is taken from:

- https://go.dev/blog/using-go-modules

The minimum Go version is currently pinned, both in the `manifest.toml`
and `go.mod`, at the latest stable major and minor version. It should
continue to work after a `flox upgrade` assuming that there are no
breaking changes.