chore: change messages for automatic environment updates (#309)
The automatic upgrades of environments in this repo acreate commit
messages and PR titles reading "Update _manifest_ of ..." which is
confusing, if the manifest remains unchanged and the lock is updated as
a result of the upgrade.

I'm not too married to say lockfile here if there was a reason against
it but "updated manifest" seems misleading.