Update podman manifest and add test (#313)
The podman manifest has not been updated since before the catalog
release, so it doesn't have `version = 1`. It also uses the old pkg-path
syntax of an array rather than a string. Fix both of those issues.

Add a test so that the environment gets updated by CI. The test runs
quay.io/podman/hello, which is currently size 580 kB, so it should be
cheap enough to run in a test.

A condition is included in the test to only run on Linux, since setting
up a podman VM on macOS would be expensive.