fix: Don't explicitly use a socket for Postgres
Previously, the default configuration for flox/postgres uses both a socket and
a TCP connection. This was both redundant and caused Postgres not to start in
several scenarios -- primarily because the postgres socket file path was deep
in the .flox structure, which already might be deep in the filesystem. The
maximum lenght for a file socket path on MacOS is 103 characters. If it's
longer than that, PG will not start. This change simply removes the call to `-c
socket` because it's redundant with the way Postgres is already configured via
this env.

Additionally, the presence of PGHOST cause the primary go Postgres library to
panic as it tries to parse that even when the presence of TCP connection
information is available.

I actually wonder if PGHOST should be renamed to PGRUN or something else, as
it's only used to locate the `/run` directory.